// express/schemas/schema.d.ts
export interface components {
  schemas: {
    /** @description A listing page number */
    Page: number
    /** @description The unique identifier */
    Id: number
    /** @description The unique identifier of the token */
    TokenId: number
    /** @description Ethereum public address */
    PublicAddress: string
    /** @description Cryptographic nonce */
    Nonce: string
    /** @description JSON Web Token */
    JWT: string
    Name: string
    LongName: string
    Description: string
    Content: string
    /** Format: url */
    Url: string
    /** Format: email */
    Email: string
    /** @enum {string} */
    UploadType: 'FILE' | 'IMAGE'
    Tags: string
    /**
     * Format: ^\d+(.\d{1,2})?$
     * @description Price of ticket in Wei units
     */
    Price: string
    PostalCode: string
    /** Format: date */
    Date: string
    /** Format: date-time */
    DateTime: string
    Rate: number
    /** @description Version 4 UUID */
    UUID: string
    /** @enum {string} */
    UserRole: 'USER' | 'EVENTS_ORGANIZER' | 'ADMINISTRATOR'
    CreateArtist: {
      pictureIdUpload: components['schemas']['Id']
      name: components['schemas']['Name']
      description: components['schemas']['Description']
    }
    UpdateArtist: components['schemas']['CreateArtist']
// ...
    CreateEvent: {
      idSubcategory: components['schemas']['Id']
      cityName: components['schemas']['Name']
      statuteIdUpload: null | components['schemas']['Id']
      nftImageIdUpload: null | components['schemas']['Id']
      name: components['schemas']['LongName']
      tags: components['schemas']['Tags']
      description: components['schemas']['Description']
      video: null | components['schemas']['Url']
      ticketPrice: components['schemas']['Price']
      ticketCount: number
      maxTicketsPerUser: number
      location: components['schemas']['LongName']
      street: components['schemas']['Name']
      postalCode: components['schemas']['PostalCode']
      start: components['schemas']['DateTime']
      publish: components['schemas']['DateTime']
      draft: boolean
      transferable: boolean
      /** @description First image is the main image of event */
      images: components['schemas']['Id'][]
      artists: components['schemas']['Id'][]
    }
    UpdateEvent: {
      idSubcategory: components['schemas']['Id']
      cityName: components['schemas']['Name']
      statuteIdUpload: null | components['schemas']['Id']
      nftImageIdUpload: null | components['schemas']['Id']
      tags: components['schemas']['Tags']
      description: components['schemas']['Description']
      video: null | components['schemas']['Url']
      location: components['schemas']['LongName']
      street: components['schemas']['Name']
      postalCode: components['schemas']['PostalCode']
      start: components['schemas']['DateTime']
      draft: boolean
      /** @description First image is the main image of event */
      images: components['schemas']['Id'][]
      artists: components['schemas']['Id'][]
    }
// ...
    UpdateUserRole: {
      role: components['schemas']['UserRole']
    }
    UpdateUser: {
      name: components['schemas']['Name']
      surname: components['schemas']['Name']
      birthdate: components['schemas']['Date']
    }
    ListUsers: {
      idUser: components['schemas']['Id']
      publicAddress: components['schemas']['PublicAddress']
      email: components['schemas']['Email']
      username: components['schemas']['Name']
      role: components['schemas']['UserRole']
      active: boolean
      created: components['schemas']['DateTime']
    }[]
// ...
    CreateReview: {
      title: components['schemas']['Name']
      eventLocation: components['schemas']['LongName']
      eventDate: components['schemas']['Date']
      content: components['schemas']['Content']
      rate: components['schemas']['Rate']
    }
// ...
  }
  responses: {
    /** @description Invalid request contents */
    BadRequest: {
      content: {
        'application/json': components['schemas']['Error']
      }
    }
// ...
  }
  parameters: never
  requestBodies: never
  headers: never
  pathItems: never
}

export type $defs = Record<string, never>

export type external = Record<string, never>

export interface operations {
// ...
  CreateArtist: {
    requestBody?: {
      content: {
        'application/json': components['schemas']['CreateArtist']
      }
    }
    responses: {
      /** @description Created artist details */
      201: {
        content: {
          'application/json': components['schemas']['GetArtist']
        }
      }
      400: components['responses']['BadRequest']
      401: components['responses']['Unauthorized']
      403: components['responses']['Forbidden']
      default: components['responses']['InternalServerError']
    }
  }
// ...
  UpdateArtist: {
    parameters: {
      path: {
        /** @description The unique identifier of the artist */
        id: components['schemas']['Id']
      }
    }
    requestBody?: {
      content: {
        'application/json': components['schemas']['UpdateArtist']
      }
    }
    responses: {
      /** @description Artist updated details */
      200: {
        content: {
          'application/json': components['schemas']['GetArtist']
        }
      }
      400: components['responses']['BadRequest']
      401: components['responses']['Unauthorized']
      403: components['responses']['Forbidden']
      default: components['responses']['InternalServerError']
    }
  }
  /** @description Create artist's performance review */
  CreateReview: {
    parameters: {
      path: {
        /** @description The unique identifier of the artist */
        id: components['schemas']['Id']
      }
    }
    requestBody?: {
      content: {
        'application/json': components['schemas']['CreateReview']
      }
    }
    responses: {
      /** @description Created review details */
      201: {
        content: {
          'application/json': components['schemas']['GetReview']
        }
      }
      400: components['responses']['BadRequest']
      401: components['responses']['Unauthorized']
      429: components['responses']['TooManyRequests']
      default: components['responses']['InternalServerError']
    }
  }
// ...
  CreateEvent: {
    requestBody?: {
      content: {
        'application/json': components['schemas']['CreateEvent']
      }
    }
    responses: {
      /** @description Created event details */
      201: {
        content: {
          'application/json': components['schemas']['GetEvent']
        }
      }
      400: components['responses']['BadRequest']
      401: components['responses']['Unauthorized']
      403: components['responses']['Forbidden']
      429: components['responses']['TooManyRequests']
      default: components['responses']['InternalServerError']
    }
  }
// ...
  UpdateEvent: {
    parameters: {
      path: {
        /** @description The unique identifier of the event */
        id: components['schemas']['Id']
      }
    }
    requestBody?: {
      content: {
        'application/json': components['schemas']['UpdateEvent']
      }
    }
    responses: {
      /** @description Event updated details */
      200: {
        content: {
          'application/json': components['schemas']['GetEvent']
        }
      }
      400: components['responses']['BadRequest']
      401: components['responses']['Unauthorized']
      403: components['responses']['Forbidden']
      default: components['responses']['InternalServerError']
    }
  }
// ...
  PassiveTicketVerification: {
    parameters: {
      path: {
        /** @description The unique identifier of the event */
        id: components['schemas']['Id']
      }
    }
    requestBody?: {
      content: {
        'application/json': {
          publicAddress: components['schemas']['PublicAddress']
        }
      }
    }
    responses: {
      /** @description Passive ticket verification result */
      200: {
        content: {
          'application/json': components['schemas']['TicketVerificationResult']
        }
      }
      400: components['responses']['BadRequest']
      401: components['responses']['Unauthorized']
      403: components['responses']['Forbidden']
      default: components['responses']['InternalServerError']
    }
  }
  /** @description Subscribe for active ticket verification, long polling - 15 seconds timeout */
  ActiveTicketVerification: {
    parameters: {
      path: {
        /** @description The unique identifier of the event */
        id: components['schemas']['Id']
      }
    }
    requestBody?: {
      content: {
        'application/json': {
          uuid: components['schemas']['UUID']
        }
      }
    }
    responses: {
      /** @description Active ticket verification result */
      200: {
        content: {
          'application/json': components['schemas']['TicketVerificationResult']
        }
      }
      /** @description Long polling timeout has passed */
      204: {
        content: never
      }
      400: components['responses']['BadRequest']
      401: components['responses']['Unauthorized']
      403: components['responses']['Forbidden']
      default: components['responses']['InternalServerError']
    }
  }
// ...
  GenerateNonce: {
    requestBody?: {
      content: {
        'application/json': {
          publicAddress: components['schemas']['PublicAddress']
        }
      }
    }
    responses: {
      /** @description User nonce */
      200: {
        content: {
          'application/json': {
            nonce: components['schemas']['Nonce']
          }
        }
      }
      400: components['responses']['BadRequest']
      /** @description User not found, return sign up token */
      404: {
        content: {
          'application/json': {
            signUpToken: components['schemas']['JWT']
          }
        }
      }
      429: components['responses']['TooManyRequests']
      default: components['responses']['InternalServerError']
    }
  }
  /** @description Sign up user account */
  SignUp: {
    requestBody?: {
      content: {
        'application/json': {
          signUpToken: components['schemas']['JWT']
          email: components['schemas']['Email']
          username: components['schemas']['Name']
          name: components['schemas']['Name']
          surname: components['schemas']['Name']
          birthdate: components['schemas']['Date']
        }
      }
    }
    responses: {
      /** @description Successfully signed up user account */
      204: {
        content: never
      }
      400: components['responses']['BadRequest']
      429: components['responses']['TooManyRequests']
      default: components['responses']['InternalServerError']
    }
  }
  /** @description Activate user account */
  ActivateUser: {
    requestBody?: {
      content: {
        'application/json': {
          activationToken: components['schemas']['JWT']
        }
      }
    }
    responses: {
      /** @description Successfully activated user account */
      204: {
        content: never
      }
      400: components['responses']['BadRequest']
      429: components['responses']['TooManyRequests']
      default: components['responses']['InternalServerError']
    }
  }
  /** @description Sign in user */
  SignIn: {
    requestBody?: {
      content: {
        'application/json': {
          message: string
          /** @description ECDSA private key signature of message */
          signature: string
        }
      }
    }
    responses: {
      /** @description JWT tokens pair */
      200: {
        content: {
          'application/json': {
            authenticationToken: components['schemas']['JWT']
            refreshToken: components['schemas']['JWT']
          }
        }
      }
      400: components['responses']['BadRequest']
      429: components['responses']['TooManyRequests']
      default: components['responses']['InternalServerError']
    }
  }
// ...
  UpdateUserRole: {
    parameters: {
      path: {
        /** @description The unique identifier of the user */
        id: components['schemas']['Id']
      }
    }
    requestBody?: {
      content: {
        'application/json': components['schemas']['UpdateUserRole']
      }
    }
    responses: {
      /** @description Updated user details */
      200: {
        content: {
          'application/json': components['schemas']['GetUser']
        }
      }
      400: components['responses']['BadRequest']
      401: components['responses']['Unauthorized']
      403: components['responses']['Forbidden']
      default: components['responses']['InternalServerError']
    }
  }
// ...
  UpdateUser: {
    requestBody?: {
      content: {
        'application/json': components['schemas']['UpdateUser']
      }
    }
    responses: {
      /** @description User updated details */
      200: {
        content: {
          'application/json': components['schemas']['GetUser']
        }
      }
      400: components['responses']['BadRequest']
      401: components['responses']['Unauthorized']
      default: components['responses']['InternalServerError']
    }
  }
// ...
}